ftp.ls: Make ListingParser subclass of ListingParser
